Use uint32_t for serial bytes waiting to fix a longstanding FIXME from when USB VCP was added.
Also merges a couple of USB VCP code cleanups from betaflight.

+++ src/main/drivers/serial_softserial.c
@@ -409,7 +409,7 @@ void onSerialRxPinChange(timerCCHandlerRec_t *cbRec, captureCompare_t capture)
     }
 }
 
-uint8_t softSerialRxBytesWaiting(serialPort_t *instance)
+uint32_t softSerialRxBytesWaiting(serialPort_t *instance)
 {
     if ((instance->mode & MODE_RX) == 0) {
         return 0;
