+++ src/test/unit/telemetry_hott_unittest.cc
@@ -187,7 +187,7 @@ uint32_t millis(void) {
 
 uint32_t micros(void) { return 0; }
 
-uint8_t serialRxBytesWaiting(serialPort_t *instance) {
+uint32_t serialRxBytesWaiting(serialPort_t *instance) {
     UNUSED(instance);
     return 0;
 }
